Finished WorldObjectLocator, lots of fixes.
Changes:
Added a setWorld() to UserInterface. It needs to pass the tile map's
extent to WorldObjectLocator, but it can't take the World ref
through the constructor cause there would be a circular dependency
with the Simulation.

Changed how color mods are applied. We now render an additional sprite
in additive blend mode and change the color on that.

Added Camera::getTileViewExtent() and moved the logic out of
WorldSpriteSorter so WorldObjectLocator could reuse it.

Changed color mods to default to {0, 0, 0}, since we're now adding
a sprite ({0, 0, 0} means no color mod, since we're adding no color).

Added TileMapBase::remFloor() for interface consistency. It just calls
the clear function.

Implemented camera -> ground ray tracing in WorldObjectLocator for
mouse picking of world objects.

Fixed some "remove" behavior in the TileUpdateSystems and TileMapBase.

Fixed TileClearLayers's serialization logic.

Moved BoundingBox's function into a source file, since there's a lot now.

Added Ray struct.

Added BoundingBox::intersects(Ray).

Renamed cellExtent -> gridCellExtent in EntityLocator for consistency.

Removed "public" from some structs (structs are public by default).

Added FPoint functions to SDLHelpers.

Added Transforms::screenToWorldRay.

+++ Source/ClientLib/Config/Private/UserConfig.cpp
@@ -80,8 +80,8 @@ void UserConfig::setWindowSize(const SDL_Rect& inWindowSize)
     if ((fullscreenMode != 0) && SDL_WasInit(SDL_INIT_VIDEO)) {
         SDL_DisplayMode displayMode;
         SDL_GetDesktopDisplayMode(0, &displayMode);
-        windowSizeWidth = static_cast<float>(displayMode.w);
-        windowSizeHeight = static_cast<float>(displayMode.h);
+        windowSizeWidth = displayMode.w;
+        windowSizeHeight = displayMode.h;
     }
 }
